requirements.txt
streamlit==0.88.0
pandas==1.3.2
DateTime==4.3
numpy==1.21.2
matplotlib==3.4.3
requests==2.26.0
beautifulsoup4==4.4.1
html-table-parser==0.1.0
python-dateutil==2.8.2